+++ src/components/panels/nav/Nav.vue
@@ -194,7 +194,6 @@
               text: 'LC Marva Manual',
               click: () => {
                 const routeData = window.open('https://libgov-my.sharepoint.com/personal/pfrank_lib_loc_gov/_layouts/15/onedrive.aspx?id=%2Fpersonal%2Fpfrank%5Flib%5Floc%5Fgov%2FDocuments%2FMarva%2DManual%2DShare%2FLibrary%2Dof%2DCongress%2DMarva%2DQuartz%2DUser%2DManual%2Epdf&parent=%2Fpersonal%2Fpfrank%5Flib%5Floc%5Fgov%2FDocuments%2FMarva%2DManual%2DShare')
-                window.open(routeData.href, '_blank');
                },
               icon:"📄"
             }
@@ -606,6 +605,31 @@
               class: (this.activeProfilePosted) ? "record-posted simptip-position-bottom" : "record-unposted simptip-position-bottom",
             }
           )
+
+          menu.push(
+            {
+              text: "Open in Folio",
+              id:"send-folio",
+              ref:"folio",
+              icon: "🇫",
+              click: () => {
+                let url = "https://c2vwscf01.loc.gov/cflsops/toolkit-training-lcsg/lcap-productivity/marva/bibId/"
+                let bibId = null
+
+                // get the bibID
+                for (let rt in this.activeProfile.rt){
+                  let type = rt.split(':').slice(-1)[0]
+                  let url = this.activeProfile.rt[rt].URI
+                  if (type=='Instance'){
+                    bibId =  url.split("/")[url.split('/').length - 1]
+                  }
+                }
+                window.open(url + bibId)
+              },
+              class: (this.activeProfilePosted) ? "record-posted-folio-ok" : "record-unposted-folio-no",
+            }
+          )
+
 
 
           if (this.preferenceStore.copyMode){
@@ -1176,6 +1200,13 @@
       margin-left: 100px;
     }
 
+    .record-unposted-folio-ok{
+      pointer-events: all;
+    }
+    .record-unposted-folio-no{
+      pointer-events: none;
+      opacity: .5;
+    }
 
     .record-posted .icon{
       color: green !important;